Add files via upload
These are my files for a web site project that I had to create and manage under my own work. Some of it was unfinished as I had ran out of time to commit to a full-scale web site, so some pages only have front end and no connection to the server side portion of the project. However if given enough time, I would have been able to produce it with the connection to the server side and made a better -- more efficient web site design.